Language Translator with Watson/test_translator.py
import unittest
from translator import english_to_french, french_to_english


class TestEnglishToFrench(unittest.TestCase):
    def test1(self):
        self.assertEqual(english_to_french('I want to tell you something'),
                         'Je veux te dire quelque chose')
        self.assertEqual(english_to_french('Hello'), 'Bonjour')
        self.assertNotEqual(english_to_french('I want to tell you something'), 'Bonjour')


class TestFrenchToEnglish(unittest.TestCase):
    def test1(self):
        self.assertEqual(french_to_english('Je veux te dire quelque chose'),
                         'I want to tell you something')
        self.assertEqual(french_to_english('Bonjour'), 'Hello')
        self.assertNotEqual(french_to_english('Je veux te dire quelque chose'), 'Hello')


unittest.main()